Add fix for magento retrocompatobility
